+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Changed
 
 - `XIT ACT`: Make the action package list reorderable
+- `XIT ACT`: Add an "Allow Unfilled" option to the CX Buy action
 - `XIT SET`: Add a confirmation popup for restore from backup
 - `XIT SORT`: Add copy/paste buttons to sorting modes
 - `item-icons`: Add icons for colony-ship-related materials